Greatly simplify the POSIX handle code
This change allows hOpenSerial to produce a fully functional `Handle`
that can be used with functions such as `hWaitForInput`.  This means a
thread can block waiting for input instead of continually polling.

Only the POSIX implementation has been updated.  Someone with access to
a Windows machine should see if this can be implemented there too.